Remplate intval() par floatval() sinon les quotas (ex: bande passante) sont limites a 2147483647 (1.99 gigs), quoique, voir aussi le bug #1129

+++ bureau/class/m_quota.php
@@ -161,7 +161,7 @@ class m_quota {
   function setquota($ressource,$size) {
     global $err,$db,$cuid;
     $err->log("quota","setquota",$ressource."/".$size);
-    if (intval($size)==0) $size="0";
+    if (floatval($size)==0) $size="0";
     if ($this->disk[$ressource]) {
       // It's a disk resource, update it with shell command
       exec("/usr/lib/alternc/quota_edit $cuid $size");
